% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherency.R
\name{CoFEScoherency}
\alias{CoFEScoherency}
\title{Title}
\usage{
CoFEScoherency(
  x,
  y,
  dt = 1,
  dj = 0.25,
  low.period = 2 * dt,
  up.period = length(x) * dt,
  pad = 0,
  sigma = 1,
  wt.type = 0,
  wt.size = 5,
  ws.type = 0,
  ws.size = 5,
  n.sur = 0,
  p = 0,
  q = 0,
  Phase_diff = FALSE,
  low.fp = 32,
  up.fp = 128,
  date = NULL
)
}
\arguments{
\item{date}{}
}
\value{

}
\description{
Title
}
