% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherency_partial.R
\name{CoFESmpcoherency}
\alias{CoFESmpcoherency}
\title{Title}
\usage{
CoFESmpcoherency(
  X,
  dt = 1,
  dj = 0.25,
  low.period = 2 * dt,
  up.period = nrow(X) * dt,
  pad = 0,
  sigma = 1,
  coher.type = 0,
  index.p = 2,
  wt.type = 0,
  wt.size = 5,
  ws.type = 0,
  ws.size = 5,
  n.sur = 0,
  p = 0,
  q = 0,
  Phase_diff = FALSE,
  low.fp = 32,
  up.fp = 128,
  date = NULL
)
}
\arguments{
\item{date}{}
}
\value{

}
\description{
Title
}
