% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{CoFESWave.image}
\alias{CoFESWave.image}
\title{Title}
\usage{
CoFESWave.image(
  WT,
  my.series = 1,
  exponent = 1,
  plot.coi = TRUE,
  plot.contour = TRUE,
  siglvl = 0.1,
  col.contour = "white",
  plot.ridge = TRUE,
  lvl = 0,
  col.ridge = "black",
  color.key = "quantile",
  n.levels = 100,
  color.palette = "rainbow(n.levels, start = 0, end = .7)",
  maximum.level = NULL,
  useRaster = TRUE,
  max.contour.segments = 250000,
  plot.legend = TRUE,
  legend.params = list(width = 1.2, shrink = 0.9, mar = 5.1, n.ticks = 6, label.digits =
    1, label.format = "f", lab = NULL, lab.line = 2.5),
  label.time.axis = TRUE,
  show.date = FALSE,
  date.format = NULL,
  date.tz = NULL,
  timelab = NULL,
  timetck = 0.02,
  timetcl = 0.5,
  spec.time.axis = list(at = NULL, labels = TRUE, las = 1, hadj = NA, padj = NA),
  label.period.axis = TRUE,
  periodlab = NULL,
  periodtck = 0.02,
  periodtcl = 0.5,
  spec.period.axis = list(at = NULL, labels = TRUE, las = 1, hadj = NA, padj = NA),
  main = NULL,
  lwd = 2,
  lwd.axis = 1,
  graphics.reset = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
