% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoFESWave.R
\docType{package}
\name{CoFESWave}
\alias{CoFESWave}
\title{WaveL2E: A package for ...}
\description{
The WaveL2E package provides ? categories of important functions:
WaveL2E, plot.CoFESWaveWCO, etc.
}
\section{CoFESWave functions}{

WaveL2E
}

