% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruct.R
\name{CoFESWave.reconstruct}
\alias{CoFESWave.reconstruct}
\title{Title}
\usage{
CoFESWave.reconstruct(
  WT,
  my.series = 1,
  lvl = 0,
  only.ridge = FALSE,
  only.sig = TRUE,
  siglvl = 0.05,
  only.coi = FALSE,
  sel.period = NULL,
  sel.lower = NULL,
  sel.upper = NULL,
  rescale = TRUE,
  plot.waves = FALSE,
  plot.rec = TRUE,
  lty = 1,
  lwd = 1,
  col = 1:2,
  ylim = NULL,
  show.legend = TRUE,
  legend.coords = "topleft",
  legend.horiz = FALSE,
  legend.text = NULL,
  label.time.axis = TRUE,
  show.date = FALSE,
  date.format = NULL,
  date.tz = NULL,
  timelab = NULL,
  timetck = 0.02,
  timetcl = 0.5,
  spec.time.axis = list(at = NULL, labels = TRUE, las = 1, hadj = NA, padj = NA),
  main.waves = NULL,
  main.rec = NULL,
  main = NULL,
  lwd.axis = 1,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
