% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{CoFESWave.Transform}
\alias{CoFESWave.Transform}
\title{Title}
\usage{
CoFESWave.Transform(
  my.data,
  my.series = 1,
  loess.span = 0.75,
  dt = 1,
  dj = 1/20,
  lowerPeriod = 2 * dt,
  upperPeriod = floor(nrow(my.data)/3) * dt,
  make.pval = TRUE,
  method = "white.noise",
  params = NULL,
  n.sim = 100,
  date.format = NULL,
  date.tz = NULL,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
