% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WaveCoFES.R
\name{plot.CoFESWaveWCO}
\alias{plot.CoFESWaveWCO}
\title{Title}
\usage{
\method{plot}{CoFESWaveWCO}(Coherency.Object, pE = 5, horizons.label = TRUE, low.FP = 32, up.FP = 128)
}
\arguments{
\item{up.FP}{}
}
\value{

}
\description{
Title
}
