% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WaveCoFES.R
\name{plot_CoFESWaveWCO}
\alias{plot_CoFESWaveWCO}
\title{Title}
\usage{
plot_CoFESWaveWCO(
  Coherency.Object,
  pE = 5,
  horizons.label = TRUE,
  low.FP = 32,
  up.FP = 128
)
}
\arguments{
\item{up.FP}{}
}
\value{

}
\description{
Title
}
