% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WaveCoFES.R
\name{plot_CoFESWaveWCO}
\alias{plot_CoFESWaveWCO}
\title{plot_CoFESWaveWCO}
\usage{
plot_CoFESWaveWCO(
  Coherency.Object,
  pE = 5,
  horizons.label = TRUE,
  low.FP = 32,
  up.FP = 128
)
}
\arguments{
\item{Coherency.Object}{output of function CoFEScoherency or CoFESmpcoherency}

\item{pE}{a constant to enhance the quality of picture}

\item{horizons.label}{Label the horizontal axis with labels (instead of numbers)? Logical.

Default: \code{TRUE}.}

\item{low.FP}{lower periods used in the frequency band}

\item{up.FP}{upper periods used in the frequency band}
}
\description{
This function plots the (absolute value) of wavelet coherency or
multiple or partial wavelet coherencies of several series, coi (and levels of sigificance)
}
\examples{
\dontrun{
## The following example is adopted from Raath et al., 2020:

## Please first run the examples in the WaveL2E, CoFEScoherency and CoFESmpcoherency functions.

# ----- Computation of coherency -----
WCO<-CoFEScoherency(Data[,1],Data[,2],low.period=low.period,up.period=up.period,
low.fp = lowFP1,up.fp = upFP1, Phase_diff = TRUE, date = date1)

# -------- PLOTS  ----------
plot_CoFESWaveWCO(WCO,pE=5, horizons.label = FALSE)


# ----- Computation of Partial Coherency of Water , Energy (controlling for SPY) ---- #

index.p=2
Data_thresh <- cbind(periodic_waveL2E_2$recon_L2E$series$x.r,
                     periodic_waveL2E$recon_L2E$series$x.r,
                     periodic_waveL2E_3$recon_L2E$series$x.r)
X<-matrix(Data_thresh,T,3)
W<-CoFESmpcoherency(X, coher.type='part',index.p=index.p,
                    low.period=low.period,up.period=up.period,
                    low.fp = lowFP1,up.fp = upFP1, Phase_diff = TRUE, date = date1)

plot_CoFESWaveWCO(W,pE=5)
}


}
\author{
CoFES. Credits are also due to L. Aguiar-Conraria and M.J. Soares.
}
