% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_WaveCoFES.R
\name{plot_CoFESWaveWCO}
\alias{plot_CoFESWaveWCO}
\title{plot_CoFESWaveWCO}
\usage{
plot_CoFESWaveWCO(
  Coherency.Object,
  pE = 5,
  horizons.label = TRUE,
  low.FP = 32,
  up.FP = 128
)
}
\arguments{
\item{Coherency.Object}{output of function CoFEScoherency or CoFESmpcoherency}

\item{pE}{a constant to enhance the quality of picture}

\item{horizons.label}{Label the horizontal axis with labels (instead of numbers)? Logical.

Default: \code{TRUE}.}

\item{low.FP}{lower periods used in the frequency band}

\item{up.FP}{upper periods used in the frequency band}
}
\description{
This function plots the (absolute value) of wavelet coherency or
multiple or partial wavelet coherencies of several series, coi (and levels of sigificance)
}
\examples{
## more work
}
\author{
CoFES. Credits are also due to L. Aguiar-Conraria and M.J. Soares.
}
