% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveL2E.R
\name{WaveL2E}
\alias{WaveL2E}
\title{WaveL2E is a function}
\usage{
WaveL2E(
  x,
  date = NULL,
  block = 1,
  base_plot = TRUE,
  L2E = TRUE,
  Chi_square = TRUE
)
}
\arguments{
\item{x}{a series}

\item{date}{a date series}

\item{block}{number of observations within a block

Default: \code{1}.}

\item{base_plot}{Plot wavelet power spectrum of x? Logical

Default: \code{TRUE}.}

\item{L2E}{Apply L2E thresholding method? Logical

Default: \code{TRUE}.}

\item{Chi_square}{Apply L2E_Chi_square thresholding method? Logical

Default: \code{TRUE}.}
}
\value{
A list of class \code{"WaveL2E"} with elements of different dimensions.
%%%%%%%%%%%%%%%%%
Here is a detailed list of all elements:
#' %%%%%%%%%%%%%%%%%
\item{sig}{tmp}
\item{w}{tmp}
\item{dis0}{tmp}
\item{thresh}{tmp}
\item{qthresh}{tmp}
\item{original}{tmp}
\item{Ana_Wave}{tmp}
\item{Emp_WaveL2E}{tmp}
\item{Emp_WaveL2E_MAD}{tmp}
\item{recon_L2E}{tmp}
\item{PTV_L2E}{tmp}
\item{PSL_L2E}{tmp}
\item{recon_Chi_square}{tmp}
\item{PTV_Chi_square}{tmp}
\item{PSL_Chi_square}{tmp}
\item{date}{a date series}
}
\description{
WaveL2E is a function
}
\examples{
## WaveL2E(x)
}
\author{
CoFES.
}
