% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveL2E.R
\name{WaveL2E}
\alias{WaveL2E}
\title{Title}
\usage{
WaveL2E(
  x,
  date = NULL,
  block = 1,
  base_plot = TRUE,
  L2E = TRUE,
  Chi_square = TRUE
)
}
\arguments{
\item{Chi_square}{}
}
\value{

}
\description{
Title
}
