% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveL2E.R
\name{WaveL2E}
\alias{WaveL2E}
\title{Wavelet Analysis with L2E and L2E_chi^2 methods}
\usage{
WaveL2E(
  x,
  date = NULL,
  block = 1,
  base_plot = TRUE,
  L2E = TRUE,
  Chi_square = TRUE
)
}
\arguments{
\item{x}{a series}

\item{date}{a date series}

\item{block}{number of observations within a block

Default: \code{1}.}

\item{base_plot}{Plot wavelet power spectrum of x? Logical

Default: \code{TRUE}.}

\item{L2E}{Apply L2E thresholding method? Logical

Default: \code{TRUE}.}

\item{Chi_square}{Apply L2E_Chi_square thresholding method? Logical

Default: \code{TRUE}.}
}
\value{
A list of class \code{"WaveL2E"} with elements of different dimensions.
%%%%%%%%%%%%%%%%%
Here is a detailed list of all elements:
#' %%%%%%%%%%%%%%%%%
\item{sig}{tmp}
\item{w}{tmp}
\item{dis0}{tmp}
\item{thresh}{threshold of \code{x} with \eqn{L2E} thresholding}
\item{qthresh}{threshold of \code{x} with \eqn{L2E(\chi^2)} thresholding}
\item{original}{wavelet analysis of \code{x} without thresholding}
\item{Ana_Wave}{wavelet analysis of \code{x} with \eqn{L2E} thresholding}
\item{Emp_WaveL2E}{tmp}
\item{Emp_WaveL2E_MAD}{tmp}
\item{recon_L2E}{reconstructed series of \code{x} with \eqn{L2E} thresholding}
\item{PTV_L2E}{percentage of total volume of \code{x} with \eqn{L2E} thresholding}
\item{PSL_L2E}{percentage of significance area of \code{x} with \eqn{L2E} thresholding}
\item{recon_Chi_square}{reconstructed series of \code{x} with \eqn{L2E(\chi^2)} thresholding}
\item{PTV_Chi_square}{percentage of total volume of \code{x} with \eqn{L2E(\chi^2)} thresholding}
\item{PSL_Chi_square}{percentage of significance area of \code{x} with \eqn{L2E(\chi^2)} thresholding}
\item{date}{the corrsponding date series}
}
\description{
This is a function based on L2E and L2E_chi^2 thresholding methods.
}
\details{
WaveL2E
}
\examples{
## WaveL2E(x)
}
\author{
CoFES.
}
