% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveL2E.R
\name{WaveL2E}
\alias{WaveL2E}
\title{WaveL2E is a function}
\usage{
WaveL2E(
  x,
  date = NULL,
  block = 1,
  base_plot = TRUE,
  L2E = TRUE,
  Chi_square = TRUE
)
}
\arguments{
\item{x}{a series}

\item{date}{a date}

\item{block}{number of what}

\item{base_plot}{TRUE if plot}

\item{L2E}{TRUE if L2E}

\item{Chi_square}{TRUE if L2E_chi_square}
}
\value{
a "whatwhat" object
}
\description{
WaveL2E is a function
}
\examples{
## WaveL2E(x)
}
